New package structure, keep backward compatibility
