+++ spiffe/spiffeid/spiffeid_test.go
@@ -1,7 +1,6 @@
 package spiffeid
 
 import (
-	"crypto/x509"
 	"net/url"
 	"testing"
 
@@ -224,37 +223,4 @@ func TestNormalizeURI(t *testing.T) {
 		assert.NoError(t, err)
 		assert.Equal(t, want, out)
 	})
-}
-
-// TestGetIDsFromCertificate covers invalid URI count.
-// Other conditions are already covered by other tests in the package.
-func TestGetIDsFromCertificate(t *testing.T) {
-	tests := []struct {
-		expectedError string
-		giveURIs      []*url.URL
-	}{
-		{
-			expectedError: "peer certificate contains no URI SAN",
-		},
-		{
-			expectedError: "peer certificate contains more than one URI SAN",
-			giveURIs: []*url.URL{
-				&url.URL{Scheme: "https", Host: "example.com"},
-				&url.URL{Scheme: "spiffe", Host: "example.net"},
-			},
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.expectedError, func(t *testing.T) {
-			giveCert := &x509.Certificate{
-				URIs: test.giveURIs,
-			}
-
-			id, domain, err := GetIDsFromCertificate(giveCert)
-			assert.Empty(t, id)
-			assert.Empty(t, domain)
-			assert.EqualError(t, err, test.expectedError)
-		})
-	}
 }
